requirements.txt
alembic==1.11.1
anyio==3.6.2
Authlib==1.2.0
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
jose==1.0.0
Mako==1.2.4
MarkupSafe==2.1.2
mypy==1.3.0
mypy-extensions==1.0.0
passlib==1.7.4
pycparser==2.21
pydantic==1.10.8
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
sqlalchemy-stubs==0.4
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.1
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
